+++ Changes.md
@@ -3,6 +3,13 @@
 
 > Note : Python 2 support has been removed. All builds are now using Python 3.
 
+Features
+--------
+
+- Mute/Solo : Added the ability to mute and solo lights.
+  - A light can be muted by setting the `light:mute` attribute. If that attribute is not present, the light will not be muted.
+  - One or more lights can be enabled exclusively via membership in the `soloLights` set. If that set has one or more members, only the lights in that set will emit. If the set is empty, all lights emit according to their mute state. If a light is both muted and in the `soloLights` set, it will emit.
+
 Fixes
 -----